refactor GodClass
