+++ src/main/java/yahoofinance/histquotes/HistQuotesRequest.java
@@ -15,9 +15,12 @@ import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import yahoofinance.Utils;
+import yahoofinance.utils.Utils;
 import yahoofinance.YahooFinance;
-import yahoofinance.util.RedirectableRequest;
+import yahoofinance.utils.BigDecimalUtil;
+import yahoofinance.utils.CalendarUtil;
+import yahoofinance.utils.PrimitiveTypesConvertUtils;
+import yahoofinance.utils.RedirectableRequest;
 
 /**
  *
@@ -137,13 +140,13 @@ public class HistQuotesRequest {
     private HistoricalQuote parseCSVLine(String line) {
         String[] data = line.split(YahooFinance.QUOTES_CSV_DELIMITER);
         return new HistoricalQuote(this.symbol,
-                Utils.parseHistDate(data[0]),
-                Utils.getBigDecimal(data[1]),
-                Utils.getBigDecimal(data[3]),
-                Utils.getBigDecimal(data[2]),
-                Utils.getBigDecimal(data[4]),
-                Utils.getBigDecimal(data[6]),
-                Utils.getLong(data[5])
+                CalendarUtil.parseHistDate(data[0]),
+                BigDecimalUtil.getBigDecimal(data[1]),
+                BigDecimalUtil.getBigDecimal(data[3]),
+                BigDecimalUtil.getBigDecimal(data[2]),
+                BigDecimalUtil.getBigDecimal(data[4]),
+                BigDecimalUtil.getBigDecimal(data[6]),
+                PrimitiveTypesConvertUtils.getLong(data[5])
         );
     }
