+++ src/main/java/yahoofinance/quotes/csv/StockQuotesData.java
@@ -1,11 +1,13 @@
 package yahoofinance.quotes.csv;
 
 import yahoofinance.Stock;
-import yahoofinance.Utils;
 import yahoofinance.exchanges.ExchangeTimeZone;
 import yahoofinance.quotes.stock.StockDividend;
 import yahoofinance.quotes.stock.StockQuote;
 import yahoofinance.quotes.stock.StockStats;
+import yahoofinance.utils.BigDecimalUtil;
+import yahoofinance.utils.CalendarUtil;
+import yahoofinance.utils.PrimitiveTypesConvertUtils;
 
 /**
  *
@@ -31,29 +33,29 @@ public class StockQuotesData {
         String symbol = this.getValue(QuotesProperty.Symbol);
         StockQuote quote = new StockQuote(symbol);
         
-        quote.setPrice(Utils.getBigDecimal(this.getValue(QuotesProperty.LastTradePriceOnly)));
-        quote.setLastTradeSize(Utils.getLong(this.getValue(QuotesProperty.LastTradeSize)));
-        quote.setAsk(Utils.getBigDecimal(this.getValue(QuotesProperty.AskRealtime), this.getValue(QuotesProperty.Ask)));
-        quote.setAskSize(Utils.getLong(this.getValue(QuotesProperty.AskSize)));
-        quote.setBid(Utils.getBigDecimal(this.getValue(QuotesProperty.BidRealtime), this.getValue(QuotesProperty.Bid)));
-        quote.setBidSize(Utils.getLong(this.getValue(QuotesProperty.BidSize)));
-        quote.setOpen(Utils.getBigDecimal(this.getValue(QuotesProperty.Open)));
-        quote.setPreviousClose(Utils.getBigDecimal(this.getValue(QuotesProperty.PreviousClose)));
-        quote.setDayHigh(Utils.getBigDecimal(this.getValue(QuotesProperty.DaysHigh)));
-        quote.setDayLow(Utils.getBigDecimal(this.getValue(QuotesProperty.DaysLow)));
+        quote.setPrice(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.LastTradePriceOnly)));
+        quote.setLastTradeSize(PrimitiveTypesConvertUtils.getLong(this.getValue(QuotesProperty.LastTradeSize)));
+        quote.setAsk(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.AskRealtime), this.getValue(QuotesProperty.Ask)));
+        quote.setAskSize(PrimitiveTypesConvertUtils.getLong(this.getValue(QuotesProperty.AskSize)));
+        quote.setBid(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.BidRealtime), this.getValue(QuotesProperty.Bid)));
+        quote.setBidSize(PrimitiveTypesConvertUtils.getLong(this.getValue(QuotesProperty.BidSize)));
+        quote.setOpen(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.Open)));
+        quote.setPreviousClose(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.PreviousClose)));
+        quote.setDayHigh(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.DaysHigh)));
+        quote.setDayLow(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.DaysLow)));
         
         quote.setTimeZone(ExchangeTimeZone.getStockTimeZone(symbol));
         quote.setLastTradeDateStr(this.getValue(QuotesProperty.LastTradeDate));
         quote.setLastTradeTimeStr(this.getValue(QuotesProperty.LastTradeTime));
-        quote.setLastTradeTime(Utils.parseDateTime(this.getValue(QuotesProperty.LastTradeDate), this.getValue(QuotesProperty.LastTradeTime), quote.getTimeZone()));
+        quote.setLastTradeTime(CalendarUtil.parseDateTime(this.getValue(QuotesProperty.LastTradeDate), this.getValue(QuotesProperty.LastTradeTime), quote.getTimeZone()));
         
-        quote.setYearHigh(Utils.getBigDecimal(this.getValue(QuotesProperty.YearHigh)));
-        quote.setYearLow(Utils.getBigDecimal(this.getValue(QuotesProperty.YearLow)));
-        quote.setPriceAvg50(Utils.getBigDecimal(this.getValue(QuotesProperty.FiftydayMovingAverage)));
-        quote.setPriceAvg200(Utils.getBigDecimal(this.getValue(QuotesProperty.TwoHundreddayMovingAverage)));
+        quote.setYearHigh(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.YearHigh)));
+        quote.setYearLow(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.YearLow)));
+        quote.setPriceAvg50(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.FiftydayMovingAverage)));
+        quote.setPriceAvg200(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.TwoHundreddayMovingAverage)));
         
-        quote.setVolume(Utils.getLong(this.getValue(QuotesProperty.Volume)));
-        quote.setAvgVolume(Utils.getLong(this.getValue(QuotesProperty.AverageDailyVolume)));
+        quote.setVolume(PrimitiveTypesConvertUtils.getLong(this.getValue(QuotesProperty.Volume)));
+        quote.setAvgVolume(PrimitiveTypesConvertUtils.getLong(this.getValue(QuotesProperty.AverageDailyVolume)));
         
         return quote;
     }
@@ -62,28 +64,28 @@ public class StockQuotesData {
         String symbol = this.getValue(QuotesProperty.Symbol);
         StockStats stats = new StockStats(symbol);
         
-        stats.setMarketCap(Utils.getBigDecimal(this.getValue(QuotesProperty.MarketCapitalization)));
-        stats.setSharesFloat(Utils.getLong(this.getValue(QuotesProperty.SharesFloat)));
-        stats.setSharesOutstanding(Utils.getLong(this.getValue(QuotesProperty.SharesOutstanding)));
-        stats.setSharesOwned(Utils.getLong(this.getValue(QuotesProperty.SharesOwned)));
+        stats.setMarketCap(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.MarketCapitalization)));
+        stats.setSharesFloat(PrimitiveTypesConvertUtils.getLong(this.getValue(QuotesProperty.SharesFloat)));
+        stats.setSharesOutstanding(PrimitiveTypesConvertUtils.getLong(this.getValue(QuotesProperty.SharesOutstanding)));
+        stats.setSharesOwned(PrimitiveTypesConvertUtils.getLong(this.getValue(QuotesProperty.SharesOwned)));
         
-        stats.setEps(Utils.getBigDecimal(this.getValue(QuotesProperty.DilutedEPS)));
-        stats.setPe(Utils.getBigDecimal(this.getValue(QuotesProperty.PERatio)));
-        stats.setPeg(Utils.getBigDecimal(this.getValue(QuotesProperty.PEGRatio)));
+        stats.setEps(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.DilutedEPS)));
+        stats.setPe(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.PERatio)));
+        stats.setPeg(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.PEGRatio)));
         
-        stats.setEpsEstimateCurrentYear(Utils.getBigDecimal(this.getValue(QuotesProperty.EPSEstimateCurrentYear)));
-        stats.setEpsEstimateNextQuarter(Utils.getBigDecimal(this.getValue(QuotesProperty.EPSEstimateNextQuarter)));
-        stats.setEpsEstimateNextYear(Utils.getBigDecimal(this.getValue(QuotesProperty.EPSEstimateNextYear)));
+        stats.setEpsEstimateCurrentYear(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.EPSEstimateCurrentYear)));
+        stats.setEpsEstimateNextQuarter(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.EPSEstimateNextQuarter)));
+        stats.setEpsEstimateNextYear(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.EPSEstimateNextYear)));
         
-        stats.setPriceBook(Utils.getBigDecimal(this.getValue(QuotesProperty.PriceBook)));
-        stats.setPriceSales(Utils.getBigDecimal(this.getValue(QuotesProperty.PriceSales)));
-        stats.setBookValuePerShare(Utils.getBigDecimal(this.getValue(QuotesProperty.BookValuePerShare)));
+        stats.setPriceBook(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.PriceBook)));
+        stats.setPriceSales(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.PriceSales)));
+        stats.setBookValuePerShare(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.BookValuePerShare)));
         
-        stats.setOneYearTargetPrice(Utils.getBigDecimal(this.getValue(QuotesProperty.OneyrTargetPrice)));
-        stats.setEBITDA(Utils.getBigDecimal(this.getValue(QuotesProperty.EBITDA)));
-        stats.setRevenue(Utils.getBigDecimal(this.getValue(QuotesProperty.Revenue)));
+        stats.setOneYearTargetPrice(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.OneyrTargetPrice)));
+        stats.setEBITDA(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.EBITDA)));
+        stats.setRevenue(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.Revenue)));
         
-        stats.setShortRatio(Utils.getBigDecimal(this.getValue(QuotesProperty.ShortRatio)));
+        stats.setShortRatio(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.ShortRatio)));
         
         return stats;
     }
@@ -92,10 +94,10 @@ public class StockQuotesData {
         String symbol = this.getValue(QuotesProperty.Symbol);
         StockDividend dividend = new StockDividend(symbol);
         
-        dividend.setPayDate(Utils.parseDividendDate(this.getValue(QuotesProperty.DividendPayDate)));
-        dividend.setExDate(Utils.parseDividendDate(this.getValue(QuotesProperty.ExDividendDate)));
-        dividend.setAnnualYield(Utils.getBigDecimal(this.getValue(QuotesProperty.TrailingAnnualDividendYield)));
-        dividend.setAnnualYieldPercent(Utils.getBigDecimal(this.getValue(QuotesProperty.TrailingAnnualDividendYieldInPercent)));
+        dividend.setPayDate(CalendarUtil.parseDividendDate(this.getValue(QuotesProperty.DividendPayDate)));
+        dividend.setExDate(CalendarUtil.parseDividendDate(this.getValue(QuotesProperty.ExDividendDate)));
+        dividend.setAnnualYield(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.TrailingAnnualDividendYield)));
+        dividend.setAnnualYieldPercent(BigDecimalUtil.getBigDecimal(this.getValue(QuotesProperty.TrailingAnnualDividendYieldInPercent)));
         
         return dividend;
     }
@@ -104,9 +106,9 @@ public class StockQuotesData {
         String symbol = this.getValue(QuotesProperty.Symbol);
         Stock stock = new Stock(symbol);
         
-        stock.setName(Utils.getString(this.getValue(QuotesProperty.Name)));
-        stock.setCurrency(Utils.getString(this.getValue(QuotesProperty.Currency)));
-        stock.setStockExchange(Utils.getString(this.getValue(QuotesProperty.StockExchange)));
+        stock.setName(PrimitiveTypesConvertUtils.getString(this.getValue(QuotesProperty.Name)));
+        stock.setCurrency(PrimitiveTypesConvertUtils.getString(this.getValue(QuotesProperty.Currency)));
+        stock.setStockExchange(PrimitiveTypesConvertUtils.getString(this.getValue(QuotesProperty.StockExchange)));
         
         stock.setQuote(this.getQuote());
         stock.setStats(this.getStats());
