+++ src/main/java/yahoofinance/quotes/query1v7/FxQuotesQuery1V7Request.java
@@ -1,8 +1,8 @@
 package yahoofinance.quotes.query1v7;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import yahoofinance.Utils;
 import yahoofinance.quotes.fx.FxQuote;
+import yahoofinance.utils.BigDecimalUtil;
 
 import java.math.BigDecimal;
 
@@ -19,7 +19,7 @@ public class FxQuotesQuery1V7Request extends QuotesRequest<FxQuote> {
     @Override
     protected FxQuote parseJson(JsonNode node) {
         String symbol = node.get("symbol").asText();
-        BigDecimal price = Utils.getBigDecimal(node.get("regularMarketPrice").asText());
+        BigDecimal price = BigDecimalUtil.getBigDecimal(node.get("regularMarketPrice").asText());
 
         return new FxQuote(symbol, price);
     }
