+++ src/main/java/yahoofinance/quotes/query1v7/StockQuotesQuery1V7Request.java
@@ -2,14 +2,15 @@ package yahoofinance.quotes.query1v7;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import yahoofinance.Stock;
-import yahoofinance.Utils;
 import yahoofinance.exchanges.ExchangeTimeZone;
 import yahoofinance.quotes.stock.StockDividend;
 import yahoofinance.quotes.stock.StockQuote;
 import yahoofinance.quotes.stock.StockStats;
+import yahoofinance.utils.BigDecimalUtil;
+import yahoofinance.utils.CalendarUtil;
+import yahoofinance.utils.PrimitiveTypesConvertUtils;
 
 import java.math.BigDecimal;
-import java.util.Arrays;
 import java.util.TimeZone;
 
 /**
@@ -58,16 +59,16 @@ public class StockQuotesQuery1V7Request extends QuotesRequest<Stock> {
         String symbol = node.get("symbol").asText();
         StockQuote quote = new StockQuote(symbol);
 
-        quote.setPrice(Utils.getBigDecimal(getStringValue(node,"regularMarketPrice")));
+        quote.setPrice(BigDecimalUtil.getBigDecimal(getStringValue(node,"regularMarketPrice")));
         // quote.setLastTradeSize(null);
-        quote.setAsk(Utils.getBigDecimal(getStringValue(node,"ask")));
-        quote.setAskSize(Utils.getLong(getStringValue(node,"askSize")));
-        quote.setBid(Utils.getBigDecimal(getStringValue(node,"bid")));
-        quote.setBidSize(Utils.getLong(getStringValue(node,"bidSize")));
-        quote.setOpen(Utils.getBigDecimal(getStringValue(node,"regularMarketOpen")));
-        quote.setPreviousClose(Utils.getBigDecimal(getStringValue(node,"regularMarketPreviousClose")));
-        quote.setDayHigh(Utils.getBigDecimal(getStringValue(node,"regularMarketDayHigh")));
-        quote.setDayLow(Utils.getBigDecimal(getStringValue(node,"regularMarketDayLow")));
+        quote.setAsk(BigDecimalUtil.getBigDecimal(getStringValue(node,"ask")));
+        quote.setAskSize(PrimitiveTypesConvertUtils.getLong(getStringValue(node,"askSize")));
+        quote.setBid(BigDecimalUtil.getBigDecimal(getStringValue(node,"bid")));
+        quote.setBidSize(PrimitiveTypesConvertUtils.getLong(getStringValue(node,"bidSize")));
+        quote.setOpen(BigDecimalUtil.getBigDecimal(getStringValue(node,"regularMarketOpen")));
+        quote.setPreviousClose(BigDecimalUtil.getBigDecimal(getStringValue(node,"regularMarketPreviousClose")));
+        quote.setDayHigh(BigDecimalUtil.getBigDecimal(getStringValue(node,"regularMarketDayHigh")));
+        quote.setDayLow(BigDecimalUtil.getBigDecimal(getStringValue(node,"regularMarketDayLow")));
 
         if(node.has("exchangeTimezoneName")) {
             quote.setTimeZone(TimeZone.getTimeZone(node.get("exchangeTimezoneName").asText()));
@@ -76,16 +77,16 @@ public class StockQuotesQuery1V7Request extends QuotesRequest<Stock> {
         }
 
         if(node.has("regularMarketTime")) {
-            quote.setLastTradeTime(Utils.unixToCalendar(node.get("regularMarketTime").asLong()));
+            quote.setLastTradeTime(CalendarUtil.unixToCalendar(node.get("regularMarketTime").asLong()));
         }
 
-        quote.setYearHigh(Utils.getBigDecimal(getStringValue(node,"fiftyTwoWeekHigh")));
-        quote.setYearLow(Utils.getBigDecimal(getStringValue(node,"fiftyTwoWeekLow")));
-        quote.setPriceAvg50(Utils.getBigDecimal(getStringValue(node,"fiftyDayAverage")));
-        quote.setPriceAvg200(Utils.getBigDecimal(getStringValue(node,"twoHundredDayAverage")));
+        quote.setYearHigh(BigDecimalUtil.getBigDecimal(getStringValue(node,"fiftyTwoWeekHigh")));
+        quote.setYearLow(BigDecimalUtil.getBigDecimal(getStringValue(node,"fiftyTwoWeekLow")));
+        quote.setPriceAvg50(BigDecimalUtil.getBigDecimal(getStringValue(node,"fiftyDayAverage")));
+        quote.setPriceAvg200(BigDecimalUtil.getBigDecimal(getStringValue(node,"twoHundredDayAverage")));
 
-        quote.setVolume(Utils.getLong(getStringValue(node,"regularMarketVolume")));
-        quote.setAvgVolume(Utils.getLong(getStringValue(node,"averageDailyVolume3Month")));
+        quote.setVolume(PrimitiveTypesConvertUtils.getLong(getStringValue(node,"regularMarketVolume")));
+        quote.setAvgVolume(PrimitiveTypesConvertUtils.getLong(getStringValue(node,"averageDailyVolume3Month")));
 
         return quote;
     }
@@ -94,22 +95,22 @@ public class StockQuotesQuery1V7Request extends QuotesRequest<Stock> {
         String symbol = getStringValue(node,"symbol");
         StockStats stats = new StockStats(symbol);
 
-        stats.setMarketCap(Utils.getBigDecimal(getStringValue(node,"marketCap")));
+        stats.setMarketCap(BigDecimalUtil.getBigDecimal(getStringValue(node,"marketCap")));
         // stats.setSharesFloat(Utils.getLong(getStringValue(node,"sharesOutstanding")));
-        stats.setSharesOutstanding(Utils.getLong(getStringValue(node,"sharesOutstanding")));
+        stats.setSharesOutstanding(PrimitiveTypesConvertUtils.getLong(getStringValue(node,"sharesOutstanding")));
         // stats.setSharesOwned(Utils.getLong(getStringValue(node,"symbol")));
 
-        stats.setEps(Utils.getBigDecimal(getStringValue(node,"epsTrailingTwelveMonths")));
-        stats.setPe(Utils.getBigDecimal(getStringValue(node,"trailingPE")));
+        stats.setEps(BigDecimalUtil.getBigDecimal(getStringValue(node,"epsTrailingTwelveMonths")));
+        stats.setPe(BigDecimalUtil.getBigDecimal(getStringValue(node,"trailingPE")));
         // stats.setPeg(Utils.getBigDecimal(getStringValue(node,"symbol")));
 
-        stats.setEpsEstimateCurrentYear(Utils.getBigDecimal(getStringValue(node,"epsForward")));
+        stats.setEpsEstimateCurrentYear(BigDecimalUtil.getBigDecimal(getStringValue(node,"epsForward")));
         // stats.setEpsEstimateNextQuarter(Utils.getBigDecimal(getStringValue(node,"symbol")));
         // stats.setEpsEstimateNextYear(Utils.getBigDecimal(getStringValue(node,"symbol")));
 
-        stats.setPriceBook(Utils.getBigDecimal(getStringValue(node,"priceToBook")));
+        stats.setPriceBook(BigDecimalUtil.getBigDecimal(getStringValue(node,"priceToBook")));
         // stats.setPriceSales(Utils.getBigDecimal(getStringValue(node,"symbol")));
-        stats.setBookValuePerShare(Utils.getBigDecimal(getStringValue(node,"bookValue")));
+        stats.setBookValuePerShare(BigDecimalUtil.getBigDecimal(getStringValue(node,"bookValue")));
 
         // stats.setOneYearTargetPrice(Utils.getBigDecimal(getStringValue(node,"symbol")));
         // stats.setEBITDA(Utils.getBigDecimal(getStringValue(node,"symbol")));
@@ -118,7 +119,7 @@ public class StockQuotesQuery1V7Request extends QuotesRequest<Stock> {
         // stats.setShortRatio(Utils.getBigDecimal(getStringValue(node,"symbol")));
 
         if(node.has("earningsTimestamp")) {
-            stats.setEarningsAnnouncement(Utils.unixToCalendar(node.get("earningsTimestamp").asLong()));
+            stats.setEarningsAnnouncement(CalendarUtil.unixToCalendar(node.get("earningsTimestamp").asLong()));
         }
 
         return stats;
@@ -130,14 +131,14 @@ public class StockQuotesQuery1V7Request extends QuotesRequest<Stock> {
 
         if (node.has("dividendDate")) {
             long dividendTimestamp = node.get("dividendDate").asLong();
-            dividend.setPayDate(Utils.unixToCalendar(dividendTimestamp));
+            dividend.setPayDate(CalendarUtil.unixToCalendar(dividendTimestamp));
             // dividend.setExDate(Utils.unixToCalendar(node.get("dividendDate").asLong()));
         }
         if (node.has("trailingAnnualDividendRate")) {
-            dividend.setAnnualYield(Utils.getBigDecimal(this.getStringValue(node, "trailingAnnualDividendRate")));
+            dividend.setAnnualYield(BigDecimalUtil.getBigDecimal(this.getStringValue(node, "trailingAnnualDividendRate")));
         }
         if (node.has("trailingAnnualDividendYield")) {
-            BigDecimal yield = Utils.getBigDecimal(this.getStringValue(node, "trailingAnnualDividendYield"));
+            BigDecimal yield = BigDecimalUtil.getBigDecimal(this.getStringValue(node, "trailingAnnualDividendYield"));
             if (yield != null) {
                 dividend.setAnnualYieldPercent(yield.multiply(ONE_HUNDRED));
             }
